+++ src/main/scala/com/dataintuitive/viash/helpers/Circe.scala
@@ -52,12 +52,18 @@ object Circe {
     def toList = list
   }
 
-  implicit def oneOrMoreToList[A](oom: OneOrMore[A]): List[A] = oom.toList
+  implicit def oneOrMoreToList[A](oom: OneOrMore[A]): List[A] = {
+    if (oom == null) {
+      Nil
+    } else {
+      oom.toList
+    }
+  }
   implicit def listToOneOrMore[A](li: List[A]): OneOrMore[A] = More(li)
 
   
   implicit def encodeOneOrMore[A](implicit enc: Encoder[List[A]]): Encoder[OneOrMore[A]] = { 
-    oom: OneOrMore[A] => enc(oom.toList)
+    oom: OneOrMore[A] => if (oom == null) enc(Nil) else enc(oom.toList)
   }
 
   implicit def decodeOneOrMore[A](implicit da: Decoder[A], dl: Decoder[List[A]]): Decoder[OneOrMore[A]] = {